requirements/base.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
cachetools==6.0.0
celery==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.9.0
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
ftfy==6.3.1
GPUtil==1.4.0
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
imageio==2.37.0
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.3
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
lazy_loader==0.4
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.5
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.63.2
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
paddleocr==3.0.0
paddlex==3.0.0
pandas==1.5.3
pillow==11.2.1
premailer==3.10.0
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.11.5
pydantic_core==2.33.2
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.5.26
tiktoken==0.9.0
tokenizers==0.19.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
